requirements.txt
python-dotenv
langchain
tqdm
selenium
pydantic
langchain_google_genai